first part of cleanup
