+++ src/app/shared/components/filter-options-button/filter-options-button.component.ts
@@ -4,14 +4,8 @@ import { NotificationService } from '../../services/notification.service';
 import { PopupMessageType } from '../../enums/popup-message-type';
 import { Store } from '@ngrx/store';
 import { RegistrationEntityTypes } from '../../models/registrations/registration-entity-categories.model';
-import { ITSystemUsageActions } from 'src/app/store/it-system-usage/actions';
-import { ITSystemActions } from 'src/app/store/it-system/actions';
-import { ITInterfaceActions } from 'src/app/store/it-system-interfaces/actions';
-import { CompositeFilterDescriptor, FilterDescriptor, isCompositeFilterDescriptor, SortDescriptor } from '@progress/kendo-data-query';
-import { ITContractActions } from 'src/app/store/it-contract/actions';
-import { DataProcessingActions } from 'src/app/store/data-processing/actions';
-import { Actions, ofType } from '@ngrx/effects';
-import { map } from 'rxjs';
+import { SavedFilterState } from '../../models/grid/saved-filter-state.model';
+import { getApplyFilterAction, getSaveFilterAction } from '../../helpers/grid-filter.helpers';
 
 @Component({
   selector: 'app-filter-options-button',
@@ -24,8 +18,7 @@ export class FilterOptionsButtonComponent {
   constructor(
     private store: Store,
     private localStorage: StatePersistingService,
-    private notificationService: NotificationService,
-    private actions$: Actions
+    private notificationService: NotificationService
   ) {}
 
   onSaveClick() {
@@ -67,60 +60,4 @@ export class FilterOptionsButtonComponent {
     const applyAction = getApplyFilterAction(this.entityType);
     this.store.dispatch(applyAction(savedState));
   }
-}
-
-export type SavedFilterState = {
-  filter: CompositeFilterDescriptor | undefined;
-  sort: SortDescriptor[] | undefined;
-};
-
-export function getSaveFilterAction(entityType: RegistrationEntityTypes) {
-  switch (entityType) {
-    case 'it-system-usage':
-      return ITSystemUsageActions.saveITSystemUsageFilter;
-    case 'it-system':
-      return ITSystemActions.saveITSystemFilter;
-    case 'it-interface':
-      return ITInterfaceActions.saveITInterfacesFilter;
-    case 'it-contract':
-      return ITContractActions.saveITContractFilter;
-    case 'data-processing-registration':
-      return DataProcessingActions.saveDataProcessingFilter;
-    default:
-      throw `Save filter action for entity type ${entityType} not implemented`;
-  }
-}
-
-export function getApplyFilterAction(entityType: RegistrationEntityTypes) {
-  switch (entityType) {
-    case 'it-system-usage':
-      return ITSystemUsageActions.applyITSystemUsageFilter;
-    case 'it-system':
-      return ITSystemActions.applyITSystemFilter;
-    case 'it-interface':
-      return ITInterfaceActions.applyITInterfacesFilter;
-    case 'it-contract':
-      return ITContractActions.applyITContractFilter;
-    case 'data-processing-registration':
-      return DataProcessingActions.applyDataProcessingFilter;
-    default:
-      throw `Apply filter action for entity type ${entityType} not implemented`;
-  }
-}
-
-/*
-This function is used to initialize a grid filters subcription for applying the saved filter state.
-updateFilter is a function that is used to update the state in the filter component.
-*/
-export function initializeApplyFilterSubscription(actions$: Actions, entityType: RegistrationEntityTypes, columnField: string, updateFilter: (filter: FilterDescriptor | undefined) => void) {
-  actions$
-      .pipe(
-        ofType(getApplyFilterAction(entityType)),
-        map((action) => action.state.filter)
-      )
-      .subscribe((compFilter) => {
-        if (!compFilter) return;
-        const matchingFilter = compFilter.filters.find((filter) => !isCompositeFilterDescriptor(filter) && filter.field === columnField) as FilterDescriptor | undefined;
-        updateFilter(matchingFilter);
-      });
 }
