+++ src/app/shared/models/buttons/button-style.model.ts
@@ -1,0 +1,1 @@
+export type ButtonStyle = 'primary' | 'secondary' | 'tertiary';
