+++ src/app/shared/models/buttons/icon-type.ts
@@ -1,0 +1,1 @@
+export type IconType = 'export' | 'dots' | undefined;
