+++ src/app/shared/models/grid/saved-filter-state.model.ts
@@ -1,0 +1,6 @@
+import { CompositeFilterDescriptor, SortDescriptor } from "@progress/kendo-data-query";
+
+export type SavedFilterState = {
+  filter: CompositeFilterDescriptor | undefined;
+  sort: SortDescriptor[] | undefined;
+};
