+++ src/app/store/data-processing/actions.ts
@@ -10,11 +10,11 @@ import {
   APIResourceCollectionPermissionsResponseDTO,
   APIUpdateDataProcessingRegistrationRequestDTO,
 } from 'src/app/api/v2';
-import { SavedFilterState } from 'src/app/shared/components/filter-options-button/filter-options-button.component';
 import { DataProcessingRegistration } from 'src/app/shared/models/data-processing/data-processing.model';
 import { ExternalReferenceProperties } from 'src/app/shared/models/external-references/external-reference-properties.model';
 import { GridColumn } from 'src/app/shared/models/grid-column.model';
 import { GridState } from 'src/app/shared/models/grid-state.model';
+import { SavedFilterState } from 'src/app/shared/models/grid/saved-filter-state.model';
 
 export const DataProcessingActions = createActionGroup({
   source: 'DataProcessing',
